✨ Move param dec to class
The parameter decorator is now not applied on the class after the
class declaration, but instead added as class decorator itself.

This is imitated from TypeScript compilation.
It leads to more consistency between compilers and increases
compatibility of libraries that depend on this behavior.

Furthermore the `design:type` decorator is also added to each decorated
class (with value `Function`, as TSC does as well).